+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added setTumors on Patient class and setPatients on NaaccrData class.
 - Removed dependency to XZ compression since it doesn't look anyone uses it.
+- Fixed first two new lines characters in data files not honoring the new line character option.
 
 **Version 4.13**