+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## NAACCR XML Version History
 
-**Version 4.13**
+**Version 4.14**
 
 - Changed dictionary validation methods to return all errors instead of just the first one.
 - Added setTumors on Patient class and setPatients on NaaccrData class.
-- Removed dependency to XZ compression since it doesn't look anyone uses it.
+- Removed dependency to XZ compression since it does not look like anyone uses it.
 - Fixed first two new lines characters in data files not honoring the new line character option.
 
 **Version 4.13**